Revert "fix, feat: Welcome screen styles, create button component, keyboard hide"
