+++ .babelrc
@@ -5,7 +5,6 @@
   "env": {
     "development": {
       "plugins": [
-        "transform-decorators-legacy",
         "transform-react-jsx-source"
       ]
     }
