+++ package.json
@@ -24,8 +24,6 @@
     "preset": "react-native"
   },
   "dependencies": {
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "core-decorators": "^0.20.0",
     "prop-types": "^15.6.1",
     "react": "16.2.0",
     "react-native": "0.52.0",
