--- src/components/Button/styles.js
@@ -1,26 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-    buttonWrap: {
-        height: 50,
-        borderRadius: 25,
-        display: 'flex',
-        flexDirection: 'column',
-        justifyContent: 'center',
-        alignSelf: 'stretch'
-    },
-    buttonText: {
-        justifyContent: 'center',
-        textAlign: 'center',
-        color: '#fff',
-        fontSize: 17,
-        fontFamily: 'FuturaStd-Light'
-    },
-    touchableOpacity: {
-        display: 'flex',
-        justifyContent: 'center'
-    }
-});
-
-
-export default styles;
